+++ contracts/DelegateBurnable.sol
@@ -19,7 +19,7 @@ interface DelegateBurnable {
   function delegateAllowance(address owner, address spender) public view returns (uint256);
   function delegateTransferFrom(address from, address to, uint256 value, address origSender) public returns (bool);
   function delegateApprove(address spender, uint256 value, address origSender) public returns (bool);
-  function delegateIncreaseApproval(address spender, uint addedValue, address origSender) public returns (bool);
-  function delegateDecreaseApproval(address spender, uint subtractedValue, address origSender) public returns (bool);
+  function delegateIncreaseApproval(address spender, uint256 addedValue, address origSender) public returns (bool);
+  function delegateDecreaseApproval(address spender, uint256 subtractedValue, address origSender) public returns (bool);
   function delegateBurn(uint256 value, address origSender) public;
 }
